change register_view structure to remove hacky circular import workaround
