+++ pymovebank/app/__main__.py
@@ -1,7 +1,7 @@
 import panel as pn
 
 # all registered apps need to be imported to the same folder
-# as the apps dict is defined. this is because
+# as the applications is imported from. this is because
 # when the apps dict is imported, then it imports each app,
 # which registers them
 from pymovebank.app.apps import applications
