+++ pymovebank/app/apps/home.py
@@ -2,8 +2,7 @@
 # pylint: disable=wrong-import-position, ungrouped-imports, wrong-import-order
 import panel as pn
 
-from pymovebank.app import config
-from pymovebank.panel_utils import templater
+from pymovebank.panel_utils import templater, register_view
 
 # pylint: enable=wrong-import-position, ungrouped-imports, wrong-import-order
 
@@ -12,7 +11,7 @@ def _add_sections():
     pn.pane.Markdown("# Home page test!").servable()
 
 
-@config.register_view()
+@register_view()
 def view(app):
     return templater(app.template, main=[pn.pane.Markdown("# Select an App on the Left")])
