+++ pymovebank/app/apps/movie_maker_app.py
@@ -7,12 +7,10 @@ import panel as pn
 
 import param
 from panel.io.loading import start_loading_spinner, stop_loading_spinner
-from pymovebank.panel_utils import param_widget, templater
-from pymovebank.app import config
-
-# from holoviews.operation.datashader import datashade, shade, dynspread, spread
+from pymovebank.panel_utils import param_widget, templater, register_view
 
 logger = logging.getLogger(__file__)
+
 
 class MovieMaker(param.Parameterized):
 
@@ -84,7 +82,7 @@ class MovieMaker(param.Parameterized):
             stop_loading_spinner(self.view)
 
 
-@config.register_view()
+@register_view()
 def view(app):
     return templater(app.template, main=[MovieMaker().view])
